reorg source files in a subpackage
